% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract.R
\name{abstract}
\alias{abstract}
\title{Knit R Markdown to MS Word Abstract}
\usage{
abstract(...)
}
\arguments{
\item{...}{Arguments passed to \code{rmarkdown::\link{word_document_format}()}.}
}
\description{
Render Conference Abstract as a .docx using the custom .tex file and .lua filters
}
