% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_resource}
\alias{find_resource}
\title{A Helper function used for locating template files.}
\usage{
find_resource(template, file = "template.tex")
}
\arguments{
\item{template}{Name of target document template}

\item{file}{Name of the target tex template for this document}
}
\description{
Locates the custom .tex template file in the rmarkdown/templates directory
}
