% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript.R
\name{manuscript}
\alias{manuscript}
\title{Knit R Markdown to MS Word Manuscript}
\usage{
manuscript(...)
}
\arguments{
\item{...}{Arguments passed to \code{rmarkdown::\link{word_document_format}()}.}
}
\description{
Render Meeting Manuscript as a .docx using the custom .tex file and .lua filters
}
